Added VC code to main SNA class
